Fix boot sequence: migrations before health checker
- Extract Release GenServer for ordered boot: repos → migrations → seeds
- Health.Manager gracefully retries if DB not ready
- Disable auto-migrations in test (mix ecto.migrate handles it)
